move point lock related classes to standalone folder
