+++ utilities/transactions/lock/point/point_lock_tracker.cc
@@ -3,7 +3,7 @@
 // COPYING file in the root directory) and Apache 2.0 License
 // (found in the LICENSE.Apache file in the root directory).
 
-#include "utilities/transactions/lock/point_lock_tracker.h"
+#include "utilities/transactions/lock/point/point_lock_tracker.h"
 
 namespace ROCKSDB_NAMESPACE {
